requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
python-dateutil==2.8.1
python-telegram-bot==13.5
pytz==2021.1
selenium==3.141.0
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
